+++ spaceMissionManager-service/src/main/java/cz/muni/fi/services/impl/SpacecraftServiceImpl.java
@@ -24,26 +24,58 @@ public class SpacecraftServiceImpl implements SpacecraftService {
 
     @Override
     public void addSpacecraft(Spacecraft spacecraft) throws DataAccessException {
-        spacecraftDao.addSpacecraft(spacecraft);
+        if (spacecraft == null){
+            throw new IllegalArgumentException("Spacecraft must not be null.");
+        }
+        try {
+            spacecraftDao.addSpacecraft(spacecraft);
+        } catch (Exception e) {
+            throw new ServiceDataAccessException("Could not add spacecraft.", e);
+        }
     }
 
     @Override
     public void removeSpacecraft(Spacecraft spacecraft) throws DataAccessException {
-        spacecraftDao.removeSpacecraft(spacecraft);
+        if (spacecraft == null){
+            throw new IllegalArgumentException("Spacecraft must not be null.");
+        }
+        try {
+            spacecraftDao.removeSpacecraft(spacecraft);
+        } catch (Exception e) {
+            throw new ServiceDataAccessException("Could not remove spacecraft.", e);
+        }
     }
 
     @Override
     public List<Spacecraft> findAllSpacecrafts() throws DataAccessException {
-        return Collections.unmodifiableList(spacecraftDao.findAllSpacecrafts());
+        try {
+            return Collections.unmodifiableList(spacecraftDao.findAllSpacecrafts());
+        } catch (Exception e) {
+            throw new ServiceDataAccessException("Could not find spacecrafts.", e);
+        }
     }
 
     @Override
     public Spacecraft findSpacecraftById(Long id) throws DataAccessException {
-        return spacecraftDao.findSpacecraftById(id);
+        if (id == null){
+            throw new IllegalArgumentException("Id must not be null.");
+        }
+        try {
+            return spacecraftDao.findSpacecraftById(id);
+        } catch (Exception e) {
+            throw new ServiceDataAccessException("Could not find spacecraft.", e);
+        }
     }
 
     @Override
     public void updateSpacecraft(Spacecraft spacecraft) throws DataAccessException {
-        spacecraftDao.updateSpacecraft(spacecraft);
+        if (spacecraft == null){
+            throw new IllegalArgumentException("Spacecraft must not be null.");
+        }
+        try {
+            spacecraftDao.updateSpacecraft(spacecraft);
+        } catch (Exception e) {
+            throw new ServiceDataAccessException("Could not update spacecraft.", e);
+        }
     }
 }
